In progress: trying to fix event timing issue
I think that the (a) problem is that events are being added by other events. Where that causes problems is when one event triggers a listener to unregister during a cycle. A possible solution may be to "stage" new events so that they're only added at the end of a frame. Workin' on it